#64 rename:  팟 하위 도메인으로 카테고리 위치 이동
