+++ src/main/java/org/com/itpple/spot/server/domain/pot/category/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
-package org.com.itpple.spot.server.domain.category.repository;
+package org.com.itpple.spot.server.domain.pot.category.repository;
 
-import org.com.itpple.spot.server.domain.category.entity.Category;
+import org.com.itpple.spot.server.domain.pot.category.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
