+++ src/main/java/org/com/itpple/spot/server/domain/pot/dto/request/CreatePotRequest.java
@@ -7,11 +7,11 @@ import static org.com.itpple.spot.server.global.util.GeometryUtil.createPoint;
 import java.time.LocalDateTime;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
-import org.com.itpple.spot.server.global.common.constant.PotType;
 import org.com.itpple.spot.server.domain.location.dto.PointDTO;
-import org.com.itpple.spot.server.domain.category.entity.Category;
+import org.com.itpple.spot.server.domain.pot.category.entity.Category;
 import org.com.itpple.spot.server.domain.pot.entity.Pot;
 import org.com.itpple.spot.server.domain.user.entity.User;
+import org.com.itpple.spot.server.global.common.constant.PotType;
 
 public record CreatePotRequest(
         @NotNull Long categoryId,
