+++ src/main/java/org/com/itpple/spot/server/domain/pot/dto/response/GetCategoryResponse.java
@@ -2,8 +2,8 @@ package org.com.itpple.spot.server.domain.pot.dto.response;
 
 import java.util.Collection;
 import java.util.List;
+import org.com.itpple.spot.server.domain.pot.category.entity.Category;
 import org.com.itpple.spot.server.domain.pot.dto.CategoryDTO;
-import org.com.itpple.spot.server.domain.category.entity.Category;
 
 
 public record GetCategoryResponse(List<CategoryDTO> categoryList) {
