+++ src/main/java/org/com/itpple/spot/server/domain/pot/entity/Pot.java
@@ -16,11 +16,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
+import org.com.itpple.spot.server.domain.pot.category.entity.Category;
 import org.com.itpple.spot.server.domain.user.entity.User;
-import org.com.itpple.spot.server.global.common.entity.BasicDateEntity;
-import org.com.itpple.spot.server.domain.category.entity.Category;
 import org.com.itpple.spot.server.global.common.constant.PotType;
+import org.com.itpple.spot.server.global.common.entity.BasicDateEntity;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLDeleteAll;
