+++ src/main/java/org/com/itpple/spot/server/domain/pot/service/impl/PotServiceImpl.java
@@ -6,19 +6,19 @@ import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.com.itpple.spot.server.domain.pot.category.repository.CategoryRepository;
 import org.com.itpple.spot.server.domain.pot.dto.PotDTO;
 import org.com.itpple.spot.server.domain.pot.dto.SearchCondition.SearchRange;
 import org.com.itpple.spot.server.domain.pot.dto.request.CreatePotRequest;
 import org.com.itpple.spot.server.domain.pot.dto.response.CreatePotResponse;
 import org.com.itpple.spot.server.domain.pot.dto.response.GetCategoryResponse;
 import org.com.itpple.spot.server.domain.pot.dto.response.UploadImageResponse;
+import org.com.itpple.spot.server.domain.pot.repository.PotRepository;
+import org.com.itpple.spot.server.domain.pot.service.PotService;
+import org.com.itpple.spot.server.domain.user.repository.UserRepository;
 import org.com.itpple.spot.server.global.exception.CustomException;
 import org.com.itpple.spot.server.global.exception.code.ErrorCode;
-import org.com.itpple.spot.server.domain.category.repository.CategoryRepository;
-import org.com.itpple.spot.server.domain.pot.repository.PotRepository;
-import org.com.itpple.spot.server.domain.user.repository.UserRepository;
 import org.com.itpple.spot.server.global.s3.service.FileService;
-import org.com.itpple.spot.server.domain.pot.service.PotService;
 import org.com.itpple.spot.server.global.util.FileUtil;
 import org.springframework.stereotype.Service;
